Create User entity to hold user information.
Adds composer and configures its autoload, also adding PHPUnit as a
development requirement. This commit also provides a phpunit.xml
configuration to be used to run the currently available unitary test
suite.

A user name should be something easy to maintain on the file system,
hence all the validation in the name.

I am aware the way User is implemented, it is actually a value object
and not an Entity, but I find the name of the namespace more appropriate
than calling it "Value" or "ValueObject".