Move node rendering to Twig.
Creates a namespace called Legacy to contain legacy classes used by the
application, which are now auto loaded by Composer.

Rendering a project is now on a very rude template, but it
servers the purpose of separating PHP code and HTML markup. A lot of
code was refactored in the way. :)

Wilber class was created to hold some helper methods to aid refactoring
the application twards a more decoupled design, currently it only has a
factory to Twig.