Functional testing with selenium.
Provides and end-to-end test suite for the project, covering basic
functionalities like: creating projects, text nodes and image nodes.

The functional tests are not isolated, which is not good, but the
intention is to provide some crude and quick tests allowing the refactor
of the project on a more scalable solution; which will make the current
tests useless.

A new project name is created for each project to prevent file system
cache problems.

Default credentials are used every time, for every test. Current
credentials **WILL** be lost, this is shitty but so is the current
authorization system, so consider this as a feature, not a bug or
lazyness of mine.

Fixes #7.